Fix names and signature for STATLIST_GetUnitStat(Un)signed
Both functions now return signed values. 99% of the usages of the stat values are signed, as can be seen by the `sar` instruction being generated when shifting values. Most of the code does not worry about the signedness of values, hence why issues arise when reaching INT_MAX internally. This can be attained rather easily due to the fixed point representation of some stats.

A little history for those wondering why it returns the exact same thing as STATLIST_UnitGetStatValue
There have been a few reworks of the stat system over the different versions of the game.
In the versions 1.08 and below, the "base" stats and "item/skill" stats were in two seperate lists.
Then in 1.09, both lists were merged into the same system, resulting in both functions doing the same thing.
This can easily be tracked by checking the function responsible for computing the defense rate and its usage of both functions.
- 1.00 to 1.05 : UNITS_GetDefenseRate#10426. STATLIST_UnitGetStatValue#10497. STATLIST_UnitGetItemStatOrSkillStatValue#10498.
- 1.08 to 1.10f: UNITS_GetDefenseRate#10431. STATLIST_UnitGetStatValue#10519. STATLIST_UnitGetItemStatOrSkillStatValue#10520.